% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_tour.R
\name{animate_tour}
\alias{animate_tour}
\title{Animate a tour path.}
\usage{
animate_tour(
  data,
  tour_path = tourr::grand_tour(d = 2),
  display = display_scatter(),
  cols = where(is.numeric),
  ...,
  rescale = TRUE,
  sphere = FALSE
)
}
\arguments{
\item{data}{data frame containing columns to use for the tour.}

\item{tour_path}{tour path generator, defaults to 2d grand tour}

\item{display}{takes the display that is suppose to be used, defaults to
the (2D) scatter display}

\item{cols}{column selection for the tour. Specified columns must be numeric.
Uses tidyselect syntax, and defaults to all numeric columns.}

\item{...}{used to pass additional arguments \code{start}, \code{aps}, \code{fps},
and \code{max_bases} through to {tourr}, as well as \code{width} and
\code{height} to {HTMLWidgets}}

\item{rescale}{if true, rescale all variables to range [0,1]?}

\item{sphere}{if true, sphere all variables}
}
\description{
This is the function that powers all of the tour animations.  If you want
to write your own tour animation method, the best place to
start is by looking at the code for animation methods that have already
implemented in the package.
}
\examples{
animate_tour(tourr::flea, tourr::grand_tour(3), display_scatter())
}
