% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_bindings.r
\name{detourr-shiny}
\alias{detourr-shiny}
\alias{shinyRenderTour}
\alias{displayScatter2dOutput}
\alias{displayScatter3dOutput}
\title{Shiny bindings for detourr}
\usage{
shinyRenderTour(expr, env = NULL)

displayScatter2dOutput(output_id, width = "100\%", height = "400px")

displayScatter3dOutput(output_id, width = "100\%", height = "400px")
}
\arguments{
\item{expr}{an expression that generates a {detourr} widget, i.e. a call
to \code{animate_tourr}}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{output_id}{output variable to read from}

\item{width}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}
}
\description{
Output and render functions for using detourr with shiny. The output
function used must match both the display method and tour dim used,
or it will lead to strange behavour.
}
