% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_scatter.r
\name{display_scatter}
\alias{display_scatter}
\title{Animate a tour path.}
\usage{
display_scatter(center = TRUE, cache_frames = TRUE)
}
\arguments{
\item{center}{If TRUE, center the projected data to (0, 0, 0).}

\item{cache_frames}{if TRUE, cache projected data at each frame of the animation. Reduces CPU usage if the tour is cycling, but can increase memory usage considerably.}
}
\description{
Display method for a high performance 3D scatterplot.
Performance is achieved through the use of Three.js / WebGL.
}
\examples{
animate_tour(tourr::flea, -species, tourr::grand_tour(3), display_scatter())
}
