% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dependence_tour}
\alias{frozen_tour}
\alias{frozen_guided_tour}
\alias{grand_tour}
\alias{guided_section_tour}
\alias{guided_tour}
\alias{little_tour}
\alias{local_tour}
\alias{planned_tour}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tourr}{\code{\link[tourr]{dependence_tour}}, \code{\link[tourr]{frozen_guided_tour}}, \code{\link[tourr]{frozen_tour}}, \code{\link[tourr]{grand_tour}}, \code{\link[tourr]{guided_section_tour}}, \code{\link[tourr]{guided_tour}}, \code{\link[tourr]{little_tour}}, \code{\link[tourr]{local_tour}}, \code{\link[tourr:planned-tour]{planned_tour}}}
}}

