% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tour_aes}
\alias{tour_aes}
\title{Aesthetic mapping for tours}
\usage{
tour_aes(...)
}
\arguments{
\item{...}{list of name-value pairs in the form 'aesthetic = variable'. Variables are evaluated
using {tidyselect} syntax.}
}
\description{
Aesthetic mapping for tours describing how variables in the data are mapped to
visual properties of the tour animation.
}
\examples{
animate_tour(
  tourr::flea,
  -species,
  grand_tour(3),
  display_scatter(tour_aes(colour = species))
)
}
