% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_bindings.r
\name{detour-shiny}
\alias{detour-shiny}
\alias{displayScatter3dOutput}
\alias{displayScatter2dOutput}
\alias{shinyRenderDisplayScatter2d}
\alias{shinyRenderDisplayScatter3d}
\alias{display_scatter_proxy}
\alias{add_points}
\title{Shiny bindings for detourr}
\usage{
displayScatter3dOutput(output_id, width = "100\%", height = "400px")

displayScatter2dOutput(output_id, width = "100\%", height = "400px")

shinyRenderDisplayScatter2d(expr, env = parent.frame(), quoted = FALSE)

shinyRenderDisplayScatter3d(expr, env = parent.frame(), quoted = FALSE)

display_scatter_proxy(id, session = shiny::getDefaultReactiveDomain())

add_points(proxy, data, scale_attr = NULL, scale_factor = NULL)
}
\arguments{
\item{output_id}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{an expression that generates a {detourr} widget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{id}{output id of the detourr instance}

\item{session}{the Shiny session object used in the app.
Default should work for most cases}

\item{proxy}{proxy object created by \code{\link{display_scatter_proxy}}}

\item{data}{dataframe of points}

\item{scale_attr}{result of \verb{attributes(scale(<original_dataset>))}
\verb{1 / max(sqrt(rowSums(scale(<original_dataset>)^2)))}}
}
\value{
An output or render function that enables the use of the widget
within shiny applications
}
\description{
Output and render functions for using detourr with shiny. The output
function used must match both the display method and tour dim used,
or it will lead to strange behavour.

Creates a proxy object that can be used to add
or remove points to a detour instance that has
already being rendered using \code{\link{shinyRenderDisplayScatter3d}}.
To be used in Shiny apps only.
#TODO: Check if namespaced modules can work as well
}
