% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_tour.R
\name{detourr-shiny}
\alias{detourr-shiny}
\alias{detourrOutput}
\alias{renderDetourr}
\title{Shiny bindings for detourr}
\usage{
detourrOutput(output_id, widget_name = NULL, width = "100\%", height = "400px")

renderDetourr(expr, env = NULL)
}
\arguments{
\item{output_id}{output variable to read from}

\item{widget_name}{name of {detourr} widget being used.
E.g. \code{display_scatter_2d}, \code{display_scatter_3d}}

\item{width}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{expr}{an expression that generates a {detourr} widget, i.e. a call
to \code{animate_tourr}}

\item{env}{The environment in which to evaluate \code{expr}.}
}
\description{
Output and render functions for using detourr with shiny
}
