% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_scatter.r
\name{display_scatter}
\alias{display_scatter}
\title{3D Scatterplot display for tours}
\usage{
display_scatter(mapping = NULL, center = TRUE, size = 1, labels = NULL)
}
\arguments{
\item{mapping}{mapping created via \code{tour_aes()}. Currently only supports colour mapping.}

\item{center}{If TRUE, center the projected data to (0, 0, 0).}

\item{size}{point size, defaults to 1}

\item{labels}{character vector of axis labels. The default of \code{labels=NULL} will give abbreviated column names. Use \code{labels=character(0)} to remove axis labels entirely.}
}
\description{
Display method for a high performance 3D scatterplot.
Performance is achieved through the use of Three.js / WebGL.
}
\examples{
animate_tour(tourr::flea, -species, tourr::grand_tour(3), display_scatter())
}
