% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_scatter.r
\name{display_scatter}
\alias{display_scatter}
\title{3D Scatterplot display for tours}
\usage{
display_scatter(
  mapping = NULL,
  center = TRUE,
  size = 1,
  labels = TRUE,
  edges = NULL,
  axes = TRUE
)
}
\arguments{
\item{mapping}{mapping created via \code{tour_aes()}. Currently only supports colour mapping.}

\item{center}{If TRUE, center the projected data to (0, 0, 0).}

\item{size}{point size, defaults to 1}

\item{labels}{axis labels. Can be:
\itemize{
\item \code{TRUE} to use column names for axis labels
\item \code{FALSE} for no labels
\item An unnamed vector of labels with the same length as \code{cols}
\item A named vector in the form \code{c("h" = "head")}, where \code{head} is renamed to \code{h}
}}

\item{edges}{A two column data frame or numeric matrix giving indices of ends of lines.}

\item{axes}{Whether to draw axes. TRUE or FALSE}
}
\description{
Display method for a high performance 3D scatterplot.
Performance is achieved through the use of Three.js / WebGL.
}
\examples{
animate_tour(tourr::flea, -species, tourr::grand_tour(3), display_scatter())
}
