% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_scatter.r
\name{display_scatter}
\alias{display_scatter}
\title{3D Scatterplot display for tours}
\usage{
display_scatter(
  mapping = NULL,
  palette = viridisLite::viridis,
  size = 1,
  alpha = 1,
  center = TRUE,
  axes = TRUE,
  edges = NULL,
  background_colour = "white"
)
}
\arguments{
\item{mapping}{mapping created via \code{tour_aes()}. Currently supports \code{colour}
and (hover) \code{labels}.}

\item{palette}{Colour palette to use with the colour aesthetic. Can be:
\itemize{
\item A character vector of R colours. This should match the number of levels
of the colour aesthetic, or the number of bins to use for continuous colours.
\item A function which takes the number of colours to use as input and returns a
character vector of colour names and / or hex values as output.
Note that the alpha channel of \code{palette} is ignored. For opacity, please use
the \code{alpha} argument.
}}

\item{size}{point size, defaults to 1}

\item{alpha}{opacity of points ranging from 0 to 1. 0 is fully transparent
and 1 is fully opaque.}

\item{center}{If TRUE, center the projected data to (0, 0, 0).}

\item{axes}{Can be one of:
\itemize{
\item \code{TRUE} draw axes and use column names for axis labels
\item \code{FALSE} do not draw axes or labels
\item \code{NULL} draw axes with no labels
\item An unnamed vector of labels with the same length as \code{cols}
\item A named vector in the form \code{c("h" = "head")}, where \code{head} is renamed to
\code{h}
}}

\item{edges}{A two column numeric matrix giving indices of ends of lines.}

\item{background_colour}{background colour for the plot.'}
}
\description{
Display method for a high performance 3D scatterplot.
Performance is achieved through the use of Three.js / WebGL.
}
\examples{
animate_tour(tourr::flea, -species, tourr::grand_tour(3), display_scatter())
}
