% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_scatter.r
\name{display_scatter}
\alias{display_scatter}
\title{2D and 3D Scatterplot display for tours}
\usage{
display_scatter(
  mapping = NULL,
  palette = viridisLite::viridis,
  size = 1,
  alpha = 1,
  center = TRUE,
  axes = TRUE,
  edges = NULL,
  background_colour = "white",
  ...,
  paused = TRUE
)
}
\arguments{
\item{mapping}{mapping created via \code{tour_aes()}. Currently supports \code{colour}
and (hover) \code{labels}.}

\item{palette}{Colour palette to use with the colour aesthetic. Can be:
\itemize{
\item A character vector of R colours. This should match the number of levels
of the colour aesthetic, or the number of bins to use for continuous colours.
\item A function which takes the number of colours to use as input and returns a
character vector of colour names and / or hex values as output.
Note that the alpha channel of \code{palette} is ignored. For opacity, please use
the \code{alpha} argument.
}}

\item{size}{point size, defaults to 1}

\item{alpha}{opacity of points ranging from 0 to 1. 0 is fully transparent
and 1 is fully opaque.}

\item{center}{If TRUE, center the projected data to (0, 0, 0).}

\item{axes}{Can be one of:
\itemize{
\item \code{TRUE} draw axes and use column names for axis labels
\item \code{FALSE} do not draw axes or labels
\item \code{NULL} draw axes with no labels
\item An unnamed vector of labels with the same length as \code{cols}
\item A named vector in the form \code{c("h" = "head")}, where \code{head} is renamed to
\code{h}
}}

\item{edges}{A two column numeric matrix giving indices of ends of lines.}

\item{background_colour}{background colour for the plot.'}

\item{...}{used to support american spelling of colour.}

\item{paused}{whether the widget should be initialised in the 'paused' state}
}
\description{
Display method for a high performance 2D or 3D scatterplot.
Performance is achieved through the use of Three.js / WebGL, and the 2D or
3D variant is selected automatically based on the tour generator provided.
}
\details{
This display method produces an interactive scatterplot animation which
supports both 2D and 3D tours. Linked selection and filtering is also
supported using {crosstalk}. The set of interactive controls availabe are:
\itemize{
\item A timeline with a play / pause button and indicators at the position of
each basis used. The basis indicators can be hovered with the mouse to show
the index of the basis, or clicked to jump to that basis. The timeline
also allows for clicking and dragging of the scrubber to move to any
individual frame of the animation.
\item Orbit controls. For the 2D variant, this allows the projection to be
rotated by clicking and dragging from left to right. For the 3D variant,
full orbit controls are available by clicking and dragging. For both orbit
and pan controls, the scroll wheel can be used to zoom.
\item Pan controls, which work similarly to orbit controls but move the camera
laterally / vertically rather than rotating
\item Resetting of the orbit and pan controls
\item Selection and highlighting. Multiple selection is possible by using the
shift key
\item Colouring / brushing of highlighted points
}
}
\examples{
animate_tour(tourr::flea, -species, tourr::grand_tour(3), display_scatter())
}
