% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tour_aes}
\alias{tour_aes}
\title{Aesthetic mapping for tours}
\usage{
tour_aes(...)
}
\arguments{
\item{...}{list of name-value pairs in the form 'aesthetic = variable'.
Variables are evaluated using {tidyselect} syntax.}
}
\description{
Aesthetic mapping for tours describing how variables in the data are
mapped to visual properties of the tour animation.
}
\examples{
detour(tourr::flea, tour_aes(projection = -species, colour = species)) |>
  tour_path(grand_tour(3), fps = 60) |>
  display_scatter(alpha = 0.7, axes = FALSE)
}
